requirements.txt
streamlit
crewai
langchain
python-dotenv
pandas
openai
